pyproject.toml
[project]
name = "smart-timelabs-onboarding"
version = "0.1.0"
description = "Automation utilities for the Smart Timelabs onboarding project."
requires-python = ">=3.12"
dependencies = [
  "email-validator>=2.2,<3",
  "fastapi>=0.115,<1",
  "httpx>=0.27,<0.29",
  "pydantic-settings>=2.6,<3",
  "PyJWT>=2.10,<3",
  "redis>=5.2,<6",
  "uvicorn>=0.34,<1",
]

[tool.uv]
package = false
